KhanhLB update all method of form
