+++ C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/controller/form/DataFormRestController.java
@@ -16,16 +16,19 @@ import javax.validation.Valid;
 
 @RestController
 @CrossOrigin("*")
-@RequestMapping("data-form")
+@RequestMapping("/api/form")
 public class DataFormRestController {
     @Autowired
     private IDataFormService iDataFormService;
-    /*Method use: searchByContent(contentDataForm,pageable.withPage(page)) of iDataFormService to get list data from database
+    /**
+     * create by : KhanhLB
+     * Data create: 31/01/2023
+     * Method use: searchByContent(contentDataForm,pageable.withPage(page)) of iDataFormService to get list data from database
      * Use ResponseEntity to handling response, datatype: Page<DataForm>
      * Parameter:  contentDataForm,page
      * If the list returned is an empty list, return http status code : HttpStatus.NO_CONTENT
      * If the list returned is a list with data, then return http status code: HttpStatus.OK and Page<DataForm>
-     * Author: KhanhLB*/
+     * */
     @GetMapping("")
     public ResponseEntity<Page<DataForm>>searchByContent(@RequestParam(defaultValue = "")String contentDataForm,@RequestParam(defaultValue = "0") int page){
         Pageable pageable = Pageable.ofSize(5);
@@ -36,13 +39,7 @@ public class DataFormRestController {
         }
         return new ResponseEntity<>(dataFormPage,HttpStatus.OK);
     }
-    /**
-     *
-     * @param dataFormDto
-     * @param bindingResult
-     *
-     */
-    /*Method use: searchByContent(contentDataForm,pageable.withPage(page)) of iDataFormService to get list data from database
+    /**Method use: searchByContent(contentDataForm,pageable.withPage(page)) of iDataFormService to get list data from database
      * Use ResponseEntity to handling response, datatype: Page<DataForm>
      * Parameter:  contentDataForm,page
      * If the list returned is an empty list, return http status code : HttpStatus.NO_CONTENT
@@ -58,5 +55,65 @@ public class DataFormRestController {
         BeanUtils.copyProperties(dataFormDto, dataForm);
         iDataFormService.saveDataForm(dataFormDto.getContentDataForm(), dataFormDto.getUrlDataForm());
         return new ResponseEntity<>(HttpStatus.CREATED);
+    }
+    /**
+     * create by : DungND
+     * Data create: 31/01/2023
+     * funcion: findById
+     * @param 'id'
+     * @return HttpStatus.NOT_FOUND if result is not found or HttpStatus.OK is find
+     */
+    @GetMapping( "/{id}")
+    public ResponseEntity<DataForm> findByID(@PathVariable("id") long id) {
+        DataForm dataForm = iDataFormService.findByIdDataForm(id);
+        if (dataForm==null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity(dataForm, HttpStatus.OK);
+    }
+
+
+    /**
+     * create by : DungND
+     * Data create: 31/01/2023
+     * funcion: delete
+     * @param 'id'
+     * @return HttpStatus.NOT_FOUND if result is not found or HttpStatus.OK is find
+     */
+    @DeleteMapping( "/delete/{id}")
+    public ResponseEntity<DataForm> delete(@PathVariable("id") long id) {
+        DataForm dataForm = iDataFormService.findByIdDataForm(id);
+        if (dataForm==null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        iDataFormService.deleteByIdDataForm(id);
+        return new ResponseEntity(dataForm, HttpStatus.OK);
+    }
+    /**
+     * create by : DungND
+     * Data create: 31/01/2023
+     * funcion: update
+     * @param 'id'
+     * @return HttpStatus.NOT_FOUND if result is not found or HttpStatus.OK is find
+     */
+
+    @PutMapping(path = "/update/{id}")
+    public ResponseEntity<DataForm> update(@PathVariable("id") long id, @Valid @RequestBody DataFormDto dataFormDto, BindingResult bindingResult){
+        DataForm dataForm = iDataFormService.findByIdDataForm(id);
+        new DataFormDto().validate(dataFormDto,bindingResult);
+        if (bindingResult.hasErrors()) {
+            return new ResponseEntity(bindingResult.getAllErrors(), HttpStatus.NOT_MODIFIED);
+        }
+
+        if (dataForm==null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        else {
+            BeanUtils.copyProperties(dataFormDto,dataForm);
+            iDataFormService.updateDataForm(dataForm);
+            return new ResponseEntity<>(dataForm, HttpStatus.OK);
+
+        }
+
     }
 }
