--- C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/repository/form/IDetailFormRepository.java
@@ -1,8 +1,0 @@
-package com.c0722g1repobe.repository.form;
-
-import com.c0722g1repobe.entity.form.DetailForm;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface IDetailFormRepository extends JpaRepository<DetailForm,Long> {
-
-}
