--- C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/service/form/IDetailFormService.java
@@ -1,4 +1,0 @@
-package com.c0722g1repobe.service.form;
-
-public interface IDetailFormService {
-}
