--- C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/service/form/impl/DetailFormService.java
@@ -1,13 +1,0 @@
-package com.c0722g1repobe.service.form.impl;
-
-import com.c0722g1repobe.repository.form.IDetailFormRepository;
-import com.c0722g1repobe.service.form.IDetailFormService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class DetailFormService implements IDetailFormService {
-    @Autowired
-    private IDetailFormRepository iDetailFormRepository;
-public class DetailFormService {
-}
